+++ Oauth2_Implementation/Production-Grade-Oauth2-Implementation/Implementation/.env
@@ -1,0 +1,19 @@
+# Google OAuth2 Client ID
+OAUTH_CLIENT_ID=Paste OAuth2 Client Id
+
+# Google OAuth2 Client Secret
+OAUTH_CLIENT_SECRET=Paste Secret Key of OAuth2
+
+# OAuth2 Redirect URL (must match the one registered in Google Cloud Console)
+OAUTH_REDIRECT_URL=http://localhost:8080/callback
+
+# OAuth2 Scopes (space-separated if multiple)
+OAUTH_SCOPES=profile email
+
+# Redis connection details
+REDIS_ADDR=localhost:6379
+REDIS_PASSWORD=
+REDIS_DB=0
+
+# JWT Secret for signing your own tokens
+JWT_SECRET=oXyQj1C1GZGvH8c4kR6rPqkjswK2fDxV1A87bEWRYP0=
